% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_anova.R
\name{auto_anova}
\alias{auto_anova}
\title{auto anova}
\usage{
auto_anova(
  data,
  ...,
  baseline = c("mean", "median", "first_level", "user_supplied"),
  user_supplied_baseline = NULL
)
}
\arguments{
\item{data}{a data frame}

\item{...}{tidyselect specification or cols}

\item{baseline}{choose from "mean", "median", "first_level", "user_supplied". what is the baseline to compare each category to? can use the mean and median of the target variable as a global baseline}

\item{user_supplied_baseline}{if intercept is "user_supplied", can enter a numeric value}
}
\value{
data frame
}
\description{
A wrapper around lm and anova to run a regression of a continuous variable against categorical variables.
Used for determining the whether the mean of a continuous variable is statistically significant amongst different levels
of a categorical variable.
}
\details{
Columns can be inputted as unquoted names or tidyselect. Continuous and categorical variables are automatically determined.
If no character or factor column is present, the column with the lowest amount of unique values will be considered
the categorical variable.
}
