% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_boxplot.R
\name{auto_boxplot}
\alias{auto_boxplot}
\title{auto_boxplot}
\usage{
auto_boxplot(
  .data,
  continuous_outcome,
  categorical_variable,
  alpha = 0.3,
  width = 0.15,
  color_dots = "black",
  color_box = "red"
)
}
\arguments{
\item{.data}{data}

\item{continuous_outcome}{continuous y variable. unquoted column name}

\item{categorical_variable}{categorical x variable. unquoted column name}

\item{alpha}{alpha points}

\item{width}{width of jitter}

\item{color_dots}{dot color}

\item{color_box}{box color}
}
\value{
ggplot
}
\description{
auto_boxplot
}
