% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_model_accuracy.R
\name{auto_model_accuracy}
\alias{auto_model_accuracy}
\title{auto model accuracy}
\usage{
auto_model_accuracy(
  data,
  formula,
  n_folds = 4,
  as_flextable = TRUE,
  theme = "tron",
  seed = 1,
  mtry = 1,
  trees = 15L,
  min_n = 1L,
  tree_depth = 6L,
  learn_rate = 0.3,
  loss_reduction = 0,
  sample_size = 1,
  stop_iter = Inf,
  counts = FALSE,
  penalty = 0.015,
  mixture = 0.35
)
}
\arguments{
\item{data}{data frame}

\item{formula}{formula}

\item{n_folds}{number of cross validation folds}

\item{as_flextable}{if FALSE, returns a tibble}

\item{theme}{make_flextable theme}

\item{seed}{seed}

\item{penalty}{linear regularization parameter}

\item{mixture}{linear model parameter, combines l1 and l2 regularization}
}
\value{
a table
}
\description{
Runs a cross validated xgboost and regularized linear regression, and reports accuracy metrics.
Automatically determines whether the provided formula is a regression or classification.
}
