% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap_outliers.R
\name{cap_outliers}
\alias{cap_outliers}
\title{cap outliers}
\usage{
cap_outliers(x, q = 0.05, type = c("both", "upper", "lower"))
}
\arguments{
\item{x}{numeric vector}

\item{q}{decimal input to quantile function. Default q = .05 means capping at the 95% and 5% percentiles}

\item{type}{default is to do an upper and lower cap, but can choose to do only one}
}
\value{
numeric vector with capped values
}
\description{
Caps the outliers of a numeric vector using quantiles. Also returns a plot of the density showing where the caps are placed.
}
\examples{

cap_outliers(iris$Petal.Width)
}
