% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_separation.R
\name{check_dichotomous}
\alias{check_dichotomous}
\title{Check dichotomous distribution for one variable}
\usage{
check_dichotomous(var, y, DF)
}
\arguments{
\item{var}{: the variable}

\item{y}{: the variable to explain}

\item{DF}{: a dataframe}
}
\value{
check_dichotomous returns a boolean TRUE if y is a dichotomous variable of x
}
\description{
Check dichotomous distribution for one variable
}
