% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{colnames_prep}
\alias{colnames_prep}
\title{Colnames preparation}
\usage{
colnames_prep(object, type = "make.names")
}
\arguments{
\item{object}{: vector of character or dataframe}

\item{type}{(character) : use 'makes.names' to create new names in correct format for R function or 'presentation' to put it in correct way to definitive presentation}
}
\value{

}
\description{
Colnames preparation
}
