% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_separation.R
\name{complete_separation}
\alias{complete_separation}
\title{Choice of complete separation method}
\usage{
complete_separation(var, y, DF, continue = FALSE)
}
\arguments{
\item{var}{variable to be evaluate}

\item{y}{the variable to explain}

\item{DF}{the dataframe}

\item{continue}{(optional) : whether the response should be apllied to all following variables.}
}
\value{

}
\description{
Choice of complete separation method
}
