% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{data_prep_complete}
\alias{data_prep_complete}
\title{Preparing datas for logistic regression}
\usage{
data_prep_complete(DF, y = colnames(DF)[1], verbose = TRUE, keep = FALSE)
}
\arguments{
\item{DF}{dataframe : the dataframe to clean}

\item{y}{(optional) character : the variable to explain}

\item{verbose}{(optional) logical : TRUE will display some informations in the console}

\item{keep}{(optional) vector : variables part of DF's columns that should not be deleted
in the data cleaning process}
}
\value{
\code{data_prep_complete} returns a dataframe ready for a logistic regression
}
\description{
Preparing datas for logistic regression
}
