\name{description_indiv_decode}
\docType{data}
\alias{description_indiv_decode}
\title{Lengths of Major North American Rivers}
\description{Description of food habitus in french population.}
\usage{description_indiv_decode}
\format{A vector containing 141 observations.}
\source{https://www.data.gouv.fr/fr/datasets/donnees-de-consommations-et-habitudes-alimentaires-de-letude-inca-3/}
\references{Data.gouv & INCA3}
\keyword{datasets}
