% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{detect_outliers}
\alias{detect_outliers}
\title{Detect outliers}
\usage{
detect_outliers(var, k = 3, inplace = TRUE)
}
\arguments{
\item{var}{: the variable to test}

\item{k}{: factor to apply to median absolute deviation to obtain the interval}

\item{inplace}{(boolean) : whether the variable should be returned without the outliers or not}
}
\value{

}
\description{
Detect outliers
}
