% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_pred_type.R
\name{determine_pred_type}
\alias{determine_pred_type}
\title{determine pred type}
\usage{
determine_pred_type(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a string
}
\description{
determine pred type
}
\keyword{internal}
