% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_preds.R
\name{eval_preds}
\alias{eval_preds}
\title{eval_preds}
\usage{
eval_preds(.data, ..., softprob_model = NULL)
}
\arguments{
\item{.data}{dataframe as a result of \code{\link{tidy_predict}}}

\item{...}{additional metrics from \href{https://yardstick.tidymodels.org/articles/metric-types.html}{yarstick} to be calculated}

\item{softprob_model}{character name of the model used to create multiclass probabilities}
}
\value{
tibble of summarized metrics
}
\description{
Automatically evaluates predictions created by \code{\link{tidy_predict}}. No need to supply column names.
}
