% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{excel}
\alias{excel}
\title{xlsx}
\usage{
excel(table, title_sheet = NULL, file_name)
}
\arguments{
\item{table}{matrix or dataframe to write in a excel file}

\item{title_sheet}{character : title of the sheet in the excel file}
}
\value{

}
\description{
xlsx
}
