% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_formula.R
\name{f_charvec_to_formula}
\alias{f_charvec_to_formula}
\title{charvec to formula}
\usage{
f_charvec_to_formula(lhs, rhs)
}
\arguments{
\item{lhs}{lhs atomic chr vec}

\item{rhs}{rhs chr vec}
}
\value{
formula
}
\description{
takes the lhs and rhs of a formula as character vectors and outputs a formula
}
\examples{

lhs <- "Species"
rhs <- c("Petal.Width", "Custom_Var")

f_charvec_to_formula(lhs, rhs)

}
