% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{format_data}
\alias{format_data}
\title{Data formating}
\usage{
format_data(DF, type = c("plain", "no-plural"))
}
\arguments{
\item{DF}{: the dataframe to format}

\item{type}{(optional) : 'plain' for no accent nor capital letter, 'no-plural' for get rid of plural differences between factors}
}
\value{
the dataframe with no accent nor capital letters
}
\description{
Data formating
}
