% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyModels.R
\name{get_model_vars}
\alias{get_model_vars}
\title{get model vars}
\usage{
get_model_vars(data, target, ...)
}
\arguments{
\item{data}{dataframe}

\item{target}{target variable}

\item{...}{tidyselect predictor variables}
}
\value{
character vector
}
\description{
get model vars
}
\keyword{internal}
