% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_id_col.R
\name{guess_id_col}
\alias{guess_id_col}
\title{guess id col}
\usage{
guess_id_col(x, min_distinct = 3L)
}
\arguments{
\item{x}{a vector}

\item{min_distinct}{integer, allows the user to choose a minimum threshold for how many unique values are in an ID column}
}
\value{
logical
}
\description{
this function uses a series of heuristics to guess whether a column is a type of ID column.
}
