% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\title{Function Logit}
\usage{
logit(DF, y = colnames(DF)[1], response = FALSE)
}
\arguments{
\item{DF}{a dataframe}

\item{y}{(optional) a variable to explain by logit function. If kept empty, y is the first element of DF.}
}
\value{

}
\description{
Function Logit
}
