% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_selection.R
\name{multivariate_selection}
\alias{multivariate_selection}
\title{Title}
\usage{
multivariate_selection(
  DF,
  y,
  explicatives = colnames(DF)[colnames(DF) != y],
  principal_factor = NULL,
  method = "backward",
  criteria = "deviance",
  check_interactions = TRUE,
  alpha = 0.05,
  verbose = TRUE
)
}
\arguments{
\item{DF}{dataframe}

\item{y}{character}

\item{explicatives}{vector of character}

\item{principal_factor}{(optional)}

\item{method}{character}

\item{criteria}{character}

\item{check_interactions}{logical}

\item{alpha}{num}

\item{verbose}{logical}
}
\value{

}
\description{
Title
}
