% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{NA_separation}
\alias{NA_separation}
\title{Looking for complete NA separation}
\usage{
NA_separation(DF, verbose = T)
}
\arguments{
\item{DF}{dataframe}

\item{verbose}{logical : TRUE to display some information in the console}
}
\value{

}
\description{
Looking for complete NA separation
}
