% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyModels.R
\name{plot_coefs_glm}
\alias{plot_coefs_glm}
\title{plot glm coefs}
\usage{
plot_coefs_glm(glm, font = c("", "HiraKakuProN-W3"))
}
\arguments{
\item{glm}{glm}

\item{font}{font}
}
\value{
plot
}
\description{
plot glm coefs
}
\keyword{internal}
