% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_ctree.R
\name{plot_ctree}
\alias{plot_ctree}
\title{plot ctree}
\usage{
plot_ctree(ctree_obj, plot_type = c("sample", "box", "bar"))
}
\arguments{
\item{ctree_obj}{output of tidy_ctree}

\item{plot_type}{type of plot}
}
\value{
decision tree plot
}
\description{
plot ctree
}
\keyword{internal}
