% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variable_contributions.R
\name{plot_variable_contributions}
\alias{plot_variable_contributions}
\title{Plot Variable Contributions}
\usage{
plot_variable_contributions(data, target, ...)
}
\arguments{
\item{data}{dataframe}

\item{target}{target variable name}

\item{...}{tidyselect specification for explanatory variables}
}
\value{
a ggplot object
}
\description{
Return a variable importance plot and coefficient plot from a linear model. Used to easily visualize
the contributions of explanatory variables in a supervised model
}
