% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyModels.R
\name{plot_varimp_cforest}
\alias{plot_varimp_cforest}
\title{plot cforest variable importance}
\usage{
plot_varimp_cforest(cfar, font = c("", "HiraKakuProN-W3"))
}
\arguments{
\item{cfar}{cforest model}

\item{font}{font}
}
\value{
ggplot
}
\description{
plot cforest variable importance
}
\keyword{internal}
