% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_xgboost.R
\name{plot_varimp_xgboost}
\alias{plot_varimp_xgboost}
\title{Plot varimp xgboost}
\usage{
plot_varimp_xgboost(xgb, top_n = 10L, as_table = FALSE, ...)
}
\arguments{
\item{xgb}{xgb.Booster model}

\item{top_n}{top n important variables}

\item{as_table}{logical, default FALSE. If TRUE returns importances in a data frame}

\item{...}{additional porams for xgb.ggplot.importance}
}
\value{
ggplot
}
\description{
recommended parameters to control;
}
\details{
\describe{
\item{\code{top_n}}{ number of features to include in the graph}
}
}
\keyword{internal}
