% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressbar.R
\name{progressbar}
\alias{progressbar}
\title{Progressbar}
\usage{
progressbar(
  i,
  total = FALSE,
  range = 25,
  variable = NULL,
  signe = "=",
  text = "Making calculs on "
)
}
\arguments{
\item{i}{: iteration}

\item{total}{: the number of iteration}

\item{range}{(optionnal) : the length of the bar in the progressbar}

\item{variable}{: name to display on the progressbar (must change for each iteration)}

\item{signe}{(optional) the signe to display for the progressbar. Put '-' to have ---> or ">" to have ">>>>"}

\item{text}{(optional) text to display befor 'variable'}
}
\value{

}
\description{
Progressbar
}
