% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_otherwise.R
\name{select_otherwise}
\alias{select_otherwise}
\title{select otherwise}
\usage{
select_otherwise(
  .data,
  ...,
  otherwise = NULL,
  col = NULL,
  return_type = c("index", "names", "df")
)
}
\arguments{
\item{.data}{dataframe}

\item{...}{tidyselect}

\item{otherwise}{tidyselect}

\item{col}{tidyselect}

\item{return_type}{choose to return column index, names, or df. defaults to index}
}
\value{
integer vector by default. possibly data frame or character vector
}
\description{
select otherwise
}
\keyword{internal}
