% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{stability_proportion}
\alias{stability_proportion}
\title{Inclusion proportion by bootstrap}
\usage{
stability_proportion(
  DF,
  y = colnames(DF)[1],
  nbre = 30,
  criterias = c("deviance"),
  methods = c("backward", "forward"),
  keep_stability = NULL
)
}
\arguments{
\item{DF}{dataframe}

\item{y}{character : variable to explain}

\item{nbre}{integer : number of iteration in the bootstrap process}

\item{criterias}{vector : criterias to test}
}
\value{

}
\description{
Inclusion proportion by bootstrap
}
