% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{stability_select_meaningful}
\alias{stability_select_meaningful}
\title{Selection of the meaningfull variables}
\usage{
stability_select_meaningful(
  stability_object,
  cutoff = 30,
  criteria = "backward deviance"
)
}
\arguments{
\item{stability_object}{: an object with the \% of variable selection by bootstrap}

\item{cutoff}{: the cutoff to select variable}
}
\value{
a vector with the meaningful variables
}
\description{
Selection of the meaningfull variables
}
