% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_agtboost.R
\name{tidy_agtboost}
\alias{tidy_agtboost}
\title{tidy agtboost}
\usage{
tidy_agtboost(.data, formula, ...)
}
\arguments{
\item{.data}{dataframe}

\item{formula}{formula}

\item{...}{additional parameters to pass to \code{\link[agtboost]{gbt.train}}}
}
\value{
agtboost model of class \code{Rcpp_ENSEMBLE}
}
\description{
Boosted tree regression using the  \href{https://github.com/Blunde1/agtboost}{agtboost} package.
Variable importance plot is printed along with returning the model. Noise features are eliminated from the plot.
}
\details{
\href{https://arxiv.org/abs/2008.12625}{agtboost: Adaptive and Automatic Gradient Tree Boosting Computations}
}
\examples{

iris \%>\%
tidy_formula(target = Petal.Length) -> f1

iris \%>\%
 tidy_agtboost(f1)
}
