% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_anova.R
\name{tidy_anova}
\alias{tidy_anova}
\title{Tidy anova}
\usage{
tidy_anova(data, ...)
}
\arguments{
\item{data}{a data frame}

\item{cols}{tidyselect specification}
}
\value{
data frame
}
\description{
Tidy anova
}
