% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_catboost.R
\name{tidy_catboost}
\alias{tidy_catboost}
\title{tidy catboost}
\usage{
tidy_catboost(.data, target, ...)
}
\arguments{
\item{.data}{dataframe}

\item{target}{target variable}

\item{...}{tidyselect}
}
\value{
catboost model
}
\description{
tidy catboost
}
