% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyModels.R
\name{tidy_cforest}
\alias{tidy_cforest}
\title{tidy conditional inference forest}
\usage{
tidy_cforest(data, target, ..., seed = 1)
}
\arguments{
\item{data}{dataframe}

\item{target}{target variable}

\item{...}{tidyselect predictor variables}

\item{seed}{seed integer}
}
\value{
a cforest model
}
\description{
tidy conditional inference forest
}
