% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_foci.R
\name{tidy_foci}
\alias{tidy_foci}
\title{Tidy FOCI}
\usage{
tidy_foci(.data, formula, ...)
}
\arguments{
\item{.data}{data}

\item{formula}{formula}

\item{...}{other arguments to FOCI}
}
\value{
data frame
}
\description{
varaible selection with FOCI
}
\examples{

iris \%>\%
tidy_foci(Species ~ .) -> d1

d1 \%>\%
  tibble::as_tibble()
}
