% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyModels.R
\name{tidy_formula}
\alias{tidy_formula}
\title{tidy formula construction}
\usage{
tidy_formula(data, target, ...)
}
\arguments{
\item{data}{dataframe}

\item{target}{lhs}

\item{...}{tidyselect. rhs}
}
\value{
a formula
}
\description{
tidy formula construction
}
