% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyModels.R
\name{tidy_glm}
\alias{tidy_glm}
\title{tidy glm}
\usage{
tidy_glm(data, formula)
}
\arguments{
\item{data}{dataframe}

\item{formula}{formula}
}
\value{
glm model
}
\description{
Runs either a linear regression, logistic regression, or multinomial classification. The model is
automatically determined based off the nature of the target variable.
}
\examples{

# linear regression
iris \%>\%
tidy_glm(
tidy_formula(., target = Petal.Width))

# multinomial classification
iris \%>\%
tidy_glm(
tidy_formula(., target = Species))

#  logistic regression
iris \%>\%
dplyr::filter(Species != "versicolor") \%>\%
tidy_glm(
tidy_formula(., target = Species))
}
