% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyModels.R
\name{tidy_lightgbm}
\alias{tidy_lightgbm}
\title{tidy lightgbm}
\usage{
tidy_lightgbm(
  data,
  target,
  ...,
  regression_objective_fun = "regression",
  verbose = -1,
  seed = 1
)
}
\arguments{
\item{data}{dataframe}

\item{target}{target variable}

\item{...}{tidyselect predictor variables}

\item{regression_objective_fun}{objective function string}

\item{verbose}{verbose integer}

\item{seed}{seed integer}
}
\value{
lightgbm model
}
\description{
tidy lightgbm
}
