% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_predict.R
\name{tidy_predict}
\alias{tidy_predict}
\alias{tidy_predict.Rcpp_ENSEMBLE}
\alias{tidy_predict.glm}
\alias{tidy_predict.default}
\alias{tidy_predict.xgb.Booster}
\title{tidy predict}
\usage{
tidy_predict(model, newdata, form = NULL, ...)

\method{tidy_predict}{Rcpp_ENSEMBLE}(model, newdata, form = NULL, ...)

\method{tidy_predict}{glm}(model, newdata, form = NULL, ...)

\method{tidy_predict}{default}(model, newdata, form = NULL, ...)

\method{tidy_predict}{xgb.Booster}(model, newdata, form = NULL, ...)
}
\arguments{
\item{model}{model}

\item{newdata}{dataframe}

\item{form}{the formula used for the model}

\item{...}{other parameters to pass to \code{predict}}
}
\value{
dataframe
}
\description{
tidy predict
}
