% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_shap.R
\name{tidy_shap}
\alias{tidy_shap}
\title{tidy shap}
\usage{
tidy_shap(model, newdata, form = NULL, ..., top_n = 12)
}
\arguments{
\item{model}{xgboost model}

\item{newdata}{dataframe similar to model input}

\item{form}{formula used for model}

\item{...}{additional parameters for shapley value}

\item{top_n}{top n features}
}
\value{
ggplot
}
\description{
plot shapley values from an xgboost model
}
