% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{tobinary}
\alias{tobinary}
\title{Coerce a column into a binary column}
\usage{
tobinary(col, verbose = TRUE)
}
\arguments{
\item{col}{column to change into a binary column}

\item{verbose}{: TRUE to display some information in the console}
}
\value{

}
\description{
Coerce a column into a binary column
}
