% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_model.R
\name{visualize_model}
\alias{visualize_model}
\alias{visualize_model.RandomForest}
\alias{visualize_model.BinaryTree}
\alias{visualize_model.glm}
\alias{visualize_model.multinom}
\alias{visualize_model.xgb.Booster}
\alias{visualize_model.default}
\title{visualize model}
\usage{
visualize_model(model, ..., method = NULL)

\method{visualize_model}{RandomForest}(model, ..., method)

\method{visualize_model}{BinaryTree}(model, ..., method)

\method{visualize_model}{glm}(model, ..., method)

\method{visualize_model}{multinom}(model, ..., method)

\method{visualize_model}{xgb.Booster}(model, ..., method)

\method{visualize_model}{default}(model, ..., method)
}
\arguments{
\item{model}{a model}

\item{...}{additional arguments}

\item{method}{choose amongst different visualization methods}
}
\value{
a plot
}
\description{
s3 method to automatically visualize the output of of a model object. Additional arguments can be supplied for the original function.
Check the corresponding plot function documentation for any custom arguments.
}
