% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_anova.R
\name{auto_anova}
\alias{auto_anova}
\title{auto anova}
\usage{
auto_anova(data, ...)
}
\arguments{
\item{data}{a data frame}

\item{cols}{tidyselect specification}
}
\value{
data frame
}
\description{
auto anova
}
