% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_xgboost.R
\name{auto_tune_xgboost}
\alias{auto_tune_xgboost}
\title{auto_tune_xgboost}
\usage{
auto_tune_xgboost(
  .data,
  target,
  ...,
  boost_engine = c("xgboost"),
  event_level = c("first", "second"),
  n_fold = 5,
  seed = 1,
  n_iter = 100
)
}
\arguments{
\item{.data}{dataframe}

\item{target}{target}

\item{...}{explanatory vars}

\item{boost_engine}{default is xgboost. catboost and lightgbm currently not supported}

\item{event_level}{for binary classification, which factor level is the positive class. specify "second" for second level}

\item{n_fold}{integer. n folds in resamples}

\item{seed}{seed}

\item{n_iter}{n iterations for tuning}
}
\value{
workflow object
}
\description{
auto_tune_xgboost
}
