% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_xgboost.R
\name{auto_tune_xgboost}
\alias{auto_tune_xgboost}
\title{auto_tune_xgboost}
\usage{
auto_tune_xgboost(
  .data,
  formula,
  event_level = c("first", "second"),
  n_fold = 5,
  seed = 1,
  n_iter = 100
)
}
\arguments{
\item{.data}{dataframe}

\item{formula}{formula}

\item{event_level}{for binary classification, which factor level is the positive class. specify "second" for second level}

\item{n_fold}{integer. n folds in resamples}

\item{seed}{seed}

\item{n_iter}{n iterations for tuning}
}
\value{
workflow object
}
\description{
Automatically tunes an xgboost model
}
