% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap_outliers.R
\name{cap_outliers}
\alias{cap_outliers}
\title{cap_outliers}
\usage{
cap_outliers(x, q = 0.05, type = c("both", "upper", "lower"))
}
\arguments{
\item{x}{numeric vector}

\item{q}{decimal input to the quantile function to set cap. default .05 caps at the 95 and 5th percentile}

\item{type}{chr vector. where to cap: both, upper, or lower}
}
\value{
numeric vector
}
\description{
Caps the outliers of a numeric vector by percentiles. Also outputs a plot of the capped distribution
}
\examples{

cap_outliers(iris$Petal.Width)

}
