% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\title{Function Logit}
\usage{
logit(DF, y = colnames(DF)[1], auto = FALSE, verbose = NULL)
}
\arguments{
\item{DF}{a dataframe}

\item{y}{(optional) a variable to explain by logit function. If kept empty, y is the first element of DF.}

\item{auto}{: if TRUE, the decision of method regarding complete separation will be automatic}

\item{verbose}{: wheter explaination should be displayed in the console}
}
\value{

}
\description{
Function Logit
}
