% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_selection.R
\name{multivariate_selection}
\alias{multivariate_selection}
\title{Multivariate selection}
\usage{
multivariate_selection(
  DF,
  y,
  explicatives = colnames(DF)[colnames(DF) != y],
  keep = FALSE,
  principal_factor = FALSE,
  method = "backward",
  criteria = "deviance",
  check_interactions = FALSE,
  alpha = 0.05,
  verbose = TRUE
)
}
\arguments{
\item{DF}{dataframe}

\item{y}{character}

\item{explicatives}{vector of character : variables to be selected in the multivariate model}

\item{keep}{character or vector of character (optional) : variables that will be kept in the model no matter of the criteria. Every variable known in the litterature to have interaction with y or other 'keep' variable should be listed in 'keep'.}

\item{principal_factor}{(optional) : principal criteria to explain y. First variable to be selected and won't be removed from selection.}

\item{method}{character}

\item{criteria}{character}

\item{check_interactions}{logical}

\item{alpha}{num}

\item{verbose}{logical}
}
\value{

}
\description{
Multivariate selection
}
