% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_selection.R
\name{multivariate_selection}
\alias{multivariate_selection}
\title{Variable selection for a multivariate logistic regression model}
\usage{
multivariate_selection(
  DF,
  y,
  explicatives,
  keep,
  principal_factor = FALSE,
  method = "backward",
  criteria = "deviance",
  check_interactions = FALSE,
  alpha = 0.05,
  verbose = TRUE,
  delta = 7
)
}
\arguments{
\item{DF}{dataframe}

\item{y}{character : variable to explain by 'explicatives'}

\item{explicatives}{vector of character : variables to be selected in the multivariate model}

\item{keep}{character or vector (optional) : variables that will be kept in the model no matter of its statistical importance Every variable known in the litterature to have interaction with y or other 'keep' variable should be listed in 'keep'.}

\item{principal_factor}{(optional) : principal criteria to explain y. It will be th first variable to be selected and won't be removed from selection.}

\item{method}{character : can be backward, forward. You can pass other arguments such as augmented or stepwise as follow "backward stepwise augmented".}

\item{criteria}{character : the criteria that will be used to select models. Can be deviance, AIC and BIC.}

\item{check_interactions}{logical : wether interaction between the principal_factor and the other variables should be checked.}

\item{alpha}{num : the threesold that should be used to consider a test as significant}

\item{verbose}{logical : whether the details of the calculs should be displayed on the console}

\item{delta}{num : threeshold of difference in BIC or AIC to consider a model as more informative}
}
\value{

}
\description{
Variable selection for a multivariate logistic regression model
}
