% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyModels.R
\name{plot_varimp}
\alias{plot_varimp}
\title{plot cforest variable importance}
\usage{
plot_varimp(vimp)
}
\arguments{
\item{vimp}{cforest vip}
}
\value{
ggplot
}
\description{
plot cforest variable importance
}
\keyword{internal}
