% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_xgboost.R
\name{plot_varimp_xgboost}
\alias{plot_varimp_xgboost}
\title{Plot varimp xgboost}
\usage{
plot_varimp_xgboost(
  xgb,
  top_n = 10L,
  aggregate = NULL,
  as_table = FALSE,
  formula = NULL,
  measure = c("Gain", "Cover", "Frequency"),
  ...
)
}
\arguments{
\item{xgb}{xgb.Booster model}

\item{top_n}{top n important variables}

\item{aggregate}{a character vector. Predictors containing the string will be aggregated, and renamed to that string.}

\item{as_table}{logical, default FALSE. If TRUE returns importances in a data frame}

\item{formula}{formula for the model. Use to provide original names if xgboost is scrambling the names internally}

\item{measure}{choose between Gain, Cover, or Frequency for xgboost importance measure}

\item{...}{additional arguments for \code{\link[xgboost]{xgb.ggplot.importance}}}

\item{font}{font}
}
\value{
ggplot
}
\description{
recommended parameters to control;
}
\details{
\itemize{
\item{\code{top_n}}{ number of features to include in the graph}
}
}
\keyword{internal}
