% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_xgboost.R
\name{plot_varimp_xgboost}
\alias{plot_varimp_xgboost}
\title{Plot varimp xgboost}
\usage{
plot_varimp_xgboost(xgb, font = c("", "HiraKakuProN-W3"), top_n = 10, ...)
}
\arguments{
\item{xgb}{xgb.Booster model}

\item{font}{font}

\item{top_n}{top n important variables}

\item{...}{additional arguments for \code{\link[xgboost]{xgb.ggplot.importance}}}
}
\value{
ggplot
}
\description{
recommended parameters to control;
}
\details{
\itemize{
\item{\code{top_n}}{ number of features to include in the graph}
}
}
\keyword{internal}
