% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity.R
\name{propensity}
\alias{propensity}
\title{Propensity score}
\usage{
propensity(DF, y, regression)
}
\arguments{
\item{DF}{dataframe}

\item{y}{character}

\item{regression}{logical}
}
\value{

}
\description{
Propensity score
}
