% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reglog.R
\name{reglog}
\alias{reglog}
\title{Regression logistique}
\usage{
reglog(
  DF,
  y,
  explicatives = colnames(DF)[colnames(DF) != y],
  alpha = 0.05,
  verbose = TRUE,
  min_multivariate = 1
)
}
\arguments{
\item{DF}{dataframe : dataframe with y and all explicative variables.}

\item{y}{character : variable to explain.}

\item{explicatives}{character vector : variables that should explain y with logistic regression. Take all columns but y from the dataframe if kept empty.}

\item{alpha}{num : significance threeshold.}

\item{verbose}{logical : TRUE for explaination in the console}
}
\value{
rslt
}
\description{
Regression logistique
}
