% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reglog.R
\name{reglog}
\alias{reglog}
\title{Logistic Regression Function}
\usage{
reglog(DF,y,explicatives,...)
}
\arguments{
\item{DF}{dataframe, matrix or tibble that contains all explicatives variables and the variable to explain}

\item{y}{character : name of the variable to explain}

\item{explicatives}{character vector : variables that should explain y in the logistic regression.
Takes all columns but y from the dataframe if kept empty.}

\item{alpha}{num : significance threeshold used to delete non-significant variables in the multivariate model.}

\item{verbose}{logical : if TRUE, explainations are displayed in the console while running the function.}

\item{alpha_max}{num : maximum threeshold used to select the minimum multivariate variables wanted.}

\item{round}{num : number of digits to display in the final table.}

\item{keep}{all the variables that should be kept in the multivariate results}

\item{exit}{specify where do you want to display the results : console (the default), excel (in a results.xlsx file), html (using kable)}

\item{method}{the method that will be used to select variables in the multivariate model.The default method is the backward elimination. See 'details' section for more informations.}
}
\value{
reglog returns a matrix with all OR obtain from univariate model and OR obtain from the multivariate model
}
\description{
reglog is able to perform logistic regression with variable selection and gives as a result a matrix with variable names, odds-ratios,
confidence intervals and p-values of univariate and multivariate models.
}
\references{
Bursac, Z., Gauss, C.H., Williams, D.K. et al. Purposeful selection of variables in logistic regression. Source Code Biol Med 3, 17 (2008). https://doi.org/10.1186/1751-0473-3-17

Heinze G, Schemper M. A solution to the problem of separation in logistic regression. Stat Med. 2002;21(16):2409-2419. doi:10.1002/sim.1047
}
