% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reglog.R
\name{reglog}
\alias{reglog}
\title{Regression logistique}
\usage{
reglog(DF, y, explicatives = colnames(DF)[colnames(DF) != y], alpha = 0.05)
}
\arguments{
\item{y}{character : la variable a expliquer}

\item{explicatives}{character vector : variables a expliquer. Laisser vide pour prendre toute les valeurs du dataframe.}

\item{alpha}{num : seuil de significativite. Par défaut à 0,05.}

\item{datas}{dataframe :la table de donnée}
}
\value{
rslt : une matrice de résultats
}
\description{
Regression logistique
}
