% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{table1}
\alias{table1}
\title{Table 1}
\usage{
table1(
  DF,
  y,
  ynames = NULL,
  overall = TRUE,
  tests = TRUE,
  norm_test = TRUE,
  Khi2 = TRUE,
  mutation = 40,
  legend = FALSE,
  title = FALSE,
  exit = "excel"
)
}
\arguments{
\item{DF}{dataframe : datas}

\item{y}{character : columns that separte the dataframe}

\item{ynames}{vector of characters (optionnal) : names to be put in table one columns instead of y levels.}

\item{overall}{booleen (optionnal) : TRUE if an "overall" column is wanted}

\item{mutation}{numeric : number of categories to display for one variable. If more than "mutation" categories, the categories after this threeshold are wrapped into a "others" categorie.}

\item{legend}{booleen (optionnal) : TRUE if a legend is wanted under the table}

\item{title}{booleen (optionnal) : TRUE if a title is wanted}
}
\value{
results are in a dataframe
}
\description{
Table 1
}
