% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyModels.R
\name{tidy_cforest}
\alias{tidy_cforest}
\alias{plot_varimp_cforest}
\title{tidy conditional inference forest}
\usage{
tidy_cforest(data, formula, seed = 1)

plot_varimp_cforest(cfar, font = c("", "HiraKakuProN-W3"))
}
\arguments{
\item{data}{dataframe}

\item{formula}{formula}

\item{seed}{seed integer}

\item{cfar}{cforest model}

\item{font}{font}
}
\value{
a cforest model

ggplot
}
\description{
Runs a conditional inference forest.
}
\examples{

iris \%>\%
tidy_cforest(
  tidy_formula(., Petal.Width)
) -> iris_cfor

iris_cfor

iris_cfor \%>\%
plot_varimp_cforest
}
