% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyModels.R
\name{tidy_glm}
\alias{tidy_glm}
\title{tidy glm}
\usage{
tidy_glm(data, target, ...)
}
\arguments{
\item{data}{dataframe}

\item{target}{target variable}

\item{...}{tidyselect predictor variables}
}
\value{
glm model
}
\description{
tidy glm
}
