% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_xgboost.R
\name{tidy_xgboost}
\alias{tidy_xgboost}
\title{tidy xgboost}
\usage{
tidy_xgboost(.data, target, ...)
}
\arguments{
\item{.data}{dataframe}

\item{target}{target variable}

\item{...}{tidyselect}
}
\value{
xgb.Booster model
}
\description{
tidy xgboost
}
