% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_xgboost.R
\name{tidy_xgboost}
\alias{tidy_xgboost}
\title{tidy xgboost}
\usage{
tidy_xgboost(.data, formula, ...)
}
\arguments{
\item{.data}{dataframe}

\item{formula}{formula}

\item{...}{additional parameters to be passed to  \code{\link[parsnip]{boost_tree}}}
}
\value{
xgb.Booster model
}
\description{
Accepts a formula to run an xgboost model. Automatically determines whether the formula is
for classification or regression. Returns the xgboost model.
}
\examples{

iris \%>\%
tidy_xgboost(
  tidy_formula(., target= Petal.Length),
  trees = 500,
  mtry = 2
)  -> xg1

xg1 \%>\%
  visualize_model(top_n = 2)
}
